+++ e2e/support/config.js
@@ -14,7 +14,6 @@ import {
   removeDirectory,
   verifyDownloadTasks,
 } from "./commands/downloads/downloadUtils";
-import webpackConfig from "./component-webpack.config";
 import * as dbTasks from "./db_tasks";
 import { signJwt } from "./helpers/e2e-jwt-tasks";
 
@@ -206,29 +205,7 @@ const mainConfig = {
   },
 };
 
-const embeddingSdkComponentTestConfig = {
-  ...defaultConfig,
-  baseUrl: undefined, // baseUrl should not be set for component tests,
-  defaultCommandTimeout: 10000,
-  requestTimeout: 10000,
-  video: false,
-  specPattern: "e2e/test-component/scenarios/embedding-sdk/**/*.cy.spec.tsx",
-  indexHtmlFile: "e2e/support/component-index.html",
-  supportFile: "e2e/support/component-cypress.js",
-
-  reporter: mainConfig.reporter,
-  reporterOptions: mainConfig.reporterOptions,
-  retries: mainConfig.retries,
-
-  devServer: {
-    framework: "react",
-    bundler: "webpack",
-    webpackConfig: webpackConfig,
-  },
-};
-
 module.exports = {
   defaultConfig,
   mainConfig,
-  embeddingSdkComponentTestConfig,
 };
